Start splitting repository.ts: extract preferences + member notes
First step of the incremental repository.ts carve-up (audit 2026-07-28 L14).
The file was ~7,100 lines — every SQL query in the product in one module —
which made it hard to navigate and the repo's worst merge-conflict hotspot,
since nearly every feature PR appends to it.

Establishes the pattern the remaining domains will follow:

- New per-domain modules under src/storage/repository/: `preferences.ts`
  (response-style + language, issues #126/#189) and `memberNotes.ts` (#45).
  Both were already fully self-contained — they touch only `pool`, `logger` and
  the `Platform` type, and nothing else in the file calls into them except
  getMyData's one read of getResponseStyle.
- repository.ts re-exports them (`export *`), so all ~42 import sites and
  tests/repository.test.ts are untouched: callers keep importing from
  repository.js and neither know nor care which file a function lives in. That
  facade is what lets this proceed one domain per PR instead of as one
  unreviewable diff.
- A header in repository.ts tells the next contributor (usually a cold pipeline
  session) to add new queries to the domain module, and the context pack's
  module-map + DB recipe now say the same.

Pure move: the eight extracted functions are byte-identical to their previous
versions — no SQL, logic or signature change, no behaviour change. Verified by
diffing each function body against the pre-split file.
